+++ app/src/components/common/PostLink.js
@@ -2,14 +2,16 @@ import React from 'react';
 
 export default function PostLink({ title, icon, onClick, target, children }) {
 	return (
-		<button
-			title={title}
-			className={`icon postLink icon__${icon}`}
-			onClick={onClick}
-			target={target ? target : ''}
-			rel="noreferrer"
-		>
-			{children}
-		</button>
+		<div className="postLink">
+			<button
+				title={title}
+				className={`icon icon__${icon}`}
+				onClick={onClick}
+				target={target ? target : ''}
+				rel="noreferrer"
+			>
+				{children}
+			</button>
+		</div>
 	);
 }